Fixed valid hex regex and merged crypto.ts into util.ts
